Apply patch from Framco.
